requirements.txt
django
django-colorfield
djangorestframework